requirements.txt
Django
django-crispy-forms
django-environ
django-pipeline
libsass
psycopg2
pytest-django
